requirements.txt
asttokens==3.0.0
attrs==25.4.0
black==25.9.0
cairocffi==1.7.1
CairoSVG==2.8.2
certifi==2025.10.5
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.4
classify-imports==4.2.0
click==8.3.0
comm==0.2.3
cssselect2==0.8.0
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
distlib==0.4.0
dotenv==0.9.9
executing==2.2.1
filelock==3.20.0
h11==0.16.0
identify==2.6.15
idna==3.11
imageio==2.37.0
ipykernel==7.0.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.9.1
markdown-it-py==4.0.0
matplotlib-inline==0.2.1
mdurl==0.1.2
mypy==1.18.2
mypy_extensions==1.1.0
nest-asyncio==1.6.0
nodeenv==1.9.1
numpy==2.3.4
outcome==1.3.0.post0
packaging==25.0
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
pre_commit==4.3.0
prompt_toolkit==3.0.52
psutil==7.1.2
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
pygame==2.6.1
Pygments==2.19.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytokens==0.2.0
pyupgrade==3.21.0
PyYAML==6.0.3
pyzmq==27.1.0
reorder_python_imports==3.16.0
requests==2.32.5
rich==14.2.0
selenium==4.38.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.3
tinycss2==1.4.0
tokenize_rt==6.2.0
tornado==6.5.2
traitlets==5.14.3
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.5.0
virtualenv==20.35.3
wcwidth==0.2.14
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.9.0
wsproto==1.2.0
